[DEV-38] rename: components 내 login, signup 디렉토리 통합
